+++ core/src/main/towers/Minutemen.java
@@ -20,7 +20,7 @@ public class Minutemen extends Tower {
     private static final float DAMAGE = 50;
     private static final float SPEED = 50;
 
-    public Minutemen(int x, int y, float radius, float cooldown, int cost, float damage){
+    public Minutemen(int x, int y, float radius, float cooldown, float cost, float damage){
     	super(x, y, radius, NAME, cost, cooldown, SIZE, damage, PARTY);
     }
