+++ core/src/main/towers/Ranger.java
@@ -21,7 +21,7 @@ public class Ranger extends Tower {
     private static final float DAMAGE = 50;
     private static final float SPEED = 25;
 
-    public Ranger(int x, int y, float radius, float cooldown, int cost, float damage){
+    public Ranger(int x, int y, float radius, float cooldown, float cost, float damage){
     	super(x, y, radius, NAME, cost, cooldown, SIZE, damage, PARTY);
     }
